[fix] : launch appication

+++ src/main/java/fr/univrouen/delivuniv/deliveryperson/DeliveryPersonService.java
@@ -5,23 +5,19 @@ import fr.univrouen.delivuniv.dto.DeliveryPersonDto;
 import fr.univrouen.delivuniv.dto.InsertDeliveryPersonDto;
 import fr.univrouen.delivuniv.dto.SearchDeliveryPersonDto;
 import fr.univrouen.delivuniv.dto.SearchResultsDto;
+import fr.univrouen.delivuniv.exeption.RessourceNotFoundException;
+import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@AllArgsConstructor
 public class DeliveryPersonService {
-
-    private DeliveryPersonRepository deliveryPersonRepository;
-    private ModelMapper mapper;
-
-    public DeliveryPersonService(DeliveryPersonRepository deliveryPersonRepository, ModelMapper mapper) {
-        this.deliveryPersonRepository = deliveryPersonRepository;
-        this.mapper = mapper;
-    }
+    private final DeliveryPersonRepository deliveryPersonRepository;
+    private final ModelMapper mapper;
 
     public DeliveryPersonDto insert(InsertDeliveryPersonDto deliveryPerson) {
         var user = deliveryPersonRepository.save(mapper.map(deliveryPerson, DeliveryPersonEntity.class));
@@ -31,6 +27,7 @@ public class DeliveryPersonService {
     public SearchResultsDto<DeliveryPersonDto> search(SearchDeliveryPersonDto model) {
         if (model.getItemsPerPage() == null) model.setItemsPerPage(10);
         if (model.getPage() == null) model.setPage(0);
+        if (model.getSearch() == null) model.setSearch("");
         var pageable = Pageable.ofSize(model.getItemsPerPage()).withPage(model.getPage());
         if (model.getOrder() == SearchOrderEnum.ORDER_BY_CREATED_DATA_DESC) {
             return SearchResultsDto.from(deliveryPersonRepository.findAllByNameContainsAndAvailableOrderByCreatedAtDesc(model.getSearch(), model.isAvailable(), pageable)
@@ -43,16 +40,21 @@ public class DeliveryPersonService {
                 .map(deliveryPerson -> mapper.map(deliveryPerson, DeliveryPersonDto.class)));
 
     }
-    public DeliveryPersonDto update(Long id, InsertDeliveryPersonDto model) {
-        var user = deliveryPersonRepository.findById(id);
-        if (user.isPresent()) {
-            user.get().setName(model.getName());
-            user.get().setAvailable(model.isAvailable());
-            return mapper.map(deliveryPersonRepository.save(user.get()), DeliveryPersonDto.class);
+    public DeliveryPersonDto update(Long id, InsertDeliveryPersonDto model)  {
+        var deliveryPersonEntity = deliveryPersonRepository.findById(id);
+        if (deliveryPersonEntity.isEmpty()) {
+            throw new RessourceNotFoundException();
         }
-        return null;
+        deliveryPersonEntity.get().setName(model.getName());
+        deliveryPersonEntity.get().setAvailable(model.isAvailable());
+        return mapper.map(deliveryPersonRepository.save(deliveryPersonEntity.get()), DeliveryPersonDto.class);
+
     }
     public void delete(Long id) {
+        var deliveryPersonEntity = deliveryPersonRepository.findById(id);
+        if (deliveryPersonEntity.isEmpty()) {
+            throw new RessourceNotFoundException();
+        }
         deliveryPersonRepository.deleteById(id);
     }
 }
